mysql2 native connection and commands

+++ controllers/adminController.js
@@ -6,39 +6,43 @@ const VIEW_PREFIX = "admin/";
 exports.getAddProduct = (req, res, next) => {
     res.render(`${VIEW_PREFIX}add_product`, {
         pageTitle: "Add Product",
+        path: "add_product",
     });
 };
 
 // ADD NEW PRODUCT POST REQUEST
-exports.postAddProduct = (req, res, next) => {
+exports.postAddProduct = async (req, res, next) => {
     const title = req.body.title,
         price = Number(req.body.price),
         description = req.body.description,
         images = req.body.images;
     if (title && price && description && images) {
         const product = new Product(title, price, images, description);
-        product.save();
-        res.redirect("/admin/products/create");
+        let prodID = await product.save();
+        if (prodID) res.redirect(`/product/${prodID}`);
+        else res.redirect("/admin/products/create");
     } else {
         res.redirect("/admin/products");
     }
 };
 
-exports.getAdminProducts = (req, res, next) => {
+exports.getAdminProducts = async (req, res, next) => {
     res.render(`${VIEW_PREFIX}admin_products`, {
         pageTitle: "Admin Products",
-        products: Product.getAllProducts(),
+        products: await Product.getAllProducts(),
+        path: "admin_products",
     });
 };
 
-exports.getEditProduct = (req, res, next) => {
+exports.getEditProduct = async (req, res, next) => {
     const productID = req.params.id;
     if (productID) {
-        let product = Product.getProduct(productID);
+        let product = await Product.getProduct(productID);
         if (product) {
             res.render(`${VIEW_PREFIX}edit_product`, {
                 pageTitle: "Edit Product",
                 product: product,
+                path: "edit_product",
             });
         } else {
             res.redirect("/admin/products");
@@ -48,15 +52,15 @@ exports.getEditProduct = (req, res, next) => {
     }
 };
 
-exports.postEditProduct = (req, res, next) => {
+exports.postEditProduct = async (req, res, next) => {
     const productID = req.params.id;
     const title = req.body.title,
-        price = Number(req.body.price),
+        price = +req.body.price,
         description = req.body.description,
         images = req.body.images;
     if (title && price && description && images && productID) {
         const product = new Product(title, price, images, description);
-        if (Product.editProduct(productID, product)) {
+        if (await Product.editProduct(productID, product)) {
             res.redirect(`/product/${productID}`);
         } else {
             res.redirect("/admin/products");
@@ -66,9 +70,11 @@ exports.postEditProduct = (req, res, next) => {
     }
 };
 
-exports.getDeleteProduct = (req, res, next) => {
+exports.getDeleteProduct = async (req, res, next) => {
     const productID = req.params.id;
-    Product.deleteProduct(productID);
-    Cart.deleteProduct(productID);
-    res.redirect("/admin/products");
+    if (+productID) {
+        let isDeleted = await Product.deleteProduct(productID);
+        if (isDeleted) res.redirect("/admin/products");
+        else res.redirect("/");
+    }
 };
